requirements.txt
msgpack==1.0.8
packaging==24.1
pyserial==3.5
python-can==4.4.2
typing_extensions==4.12.2
wrapt==1.16.0